fixed warning
git-svn-id: svn://svn.code.sf.net/p/axtls/code/trunk@171 9a5d90b5-6617-0410-8a86-bb477d3ed2e3

+++ httpd/proc.c
@@ -1022,7 +1022,7 @@ static int sanitizehost(char *buf)
         }
 
         /* Enforce some basic URL rules... */
-        if ((isalnum(*buf) == 0 && *buf != '-' && *buf != '.') ||
+        if ((isalnum((int)(*buf)) == 0 && *buf != '-' && *buf != '.') ||
                 (*buf == '.' && *(buf+1) == '.') ||
                 (*buf == '.' && *(buf+1) == '-') ||
                 (*buf == '-' && *(buf+1) == '.'))
